web-animations/interfaces/Animatable/animate-timeline.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animatable.animate timeline tests</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animatable-animate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../testcommon.js"></script>
<script src="../../resources/keyframe-utils.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

test(function(t) {
  var div = createDiv(t);
  var anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
  assert_equals(anim.timeline, document.timeline);
}, 'Element.animate() correctly sets the Animation\'s timeline');

async_test(function(t) {
  var iframe = document.createElement('iframe');
  iframe.width = 10;
  iframe.height = 10;

  iframe.addEventListener('load', t.step_func(function() {
    var div = createDiv(t, iframe.contentDocument);
    var anim = div.animate({ opacity: [ 0, 1 ] }, 2000);
    assert_equals(anim.timeline, iframe.contentDocument.timeline);
    iframe.remove();
    t.done();
  }));

  document.body.appendChild(iframe);
}, 'Element.animate() correctly sets the Animation\'s timeline when ' +
   'triggered on an element in a different document');

</script>
</body>
